[MP] Fix more defaults for saber names and team names. Don't register cg_viewsize needlessly on client startup, update cvar as soon as it's changed

+++ codemp/cgame/cg_consolecmds.c
@@ -33,7 +33,7 @@ Keybinding command
 =================
 */
 static void CG_SizeUp_f (void) {
-	trap->Cvar_Set("cg_viewsize", va("%i",(int)(cg_viewsize.integer+10)));
+	trap->Cvar_Set( "cg_viewsize", va( "%i", cg_viewsize.integer + 10 ) );
 }
 
 /*
@@ -44,7 +44,7 @@ Keybinding command
 =================
 */
 static void CG_SizeDown_f (void) {
-	trap->Cvar_Set("cg_viewsize", va("%i",(int)(cg_viewsize.integer-10)));
+	trap->Cvar_Set( "cg_viewsize", va( "%i", cg_viewsize.integer - 10 ) );
 }
 
 /*
